Added Support for Bool Flags (more below)
I added support for boolean flags, but this turns out to be a potentially confusing mechanism for the user.. Let me explain:

If the argument is present with no value, then the opposite of the default value should be used, if one is present.
For example, say there is an argument called "--no-cache", with a default value of False.
- When we don't pass the argument, the value should be False.
- When we pass the argument, the value should be True.
- When we pass the argument with a value, ex: "--no-cache true" or "--no-cache false", the given value should be used

I'm not sure if this is the best way to go, but we'll set it like that for now.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>