[BUG]: Parsing multiple containers (read desc)
Whenever we try to parse multiple containers, and the default value has the same number of elements
as the number of instances we are trying to parse, there is a bug whereby the parsed value for each
instance `i` is set to `default_value[i]` instead.

This might be because of something in postprocessing in the field_wrapper, or duplicate if needed, etc.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>