Really Cool decoding functions (read desc!)
Added an absolute Orgy of functional programming and Typing in the serialization.py file.
Now, whenever a dataclass (let's call it T) uses JsonSerializable as a base class, it makes it possible
to decode not just the dataclass itself, but also any attribute that is marked as:
- List[T]
- Optional[T]
- List[List[T]]
- List[Optional[T]]
- List[List[Optional[T]]]

So basically, we add a lot of functionality for free whenever the JsonSerializable base is used.

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>