Making more progress (need to fix conflicts)
The FieldWrapper is created for the subgroups arg.
I now create a new DataclassWrapper and graft add it as a child of the
DataclassWrapper that contains the subgroup field.

The current issue seems to be that the ConflictResolver is seeing both
the FieldWrapper and the DataclassWrapper (I Think) as pointing to the
same destination!

Signed-off-by: Fabrice Normandin <fabrice.normandin@gmail.com>