Fix is_optional(Literal[None])
Without this, a field with a type like Literal[None, "a", "b"] may be considered required on some code paths.
In particular, when specifying a default for an argument group.
Note that this bug is difficult to work around because None | Literal["a", "b"] and Optional[Literal["a", "b"]] result in different (worse) parsing behavior.